Pull out the BSShader code from textures.
