+++ io_scene_nif/modules/property/texture/texture_export.py
@@ -39,6 +39,7 @@
 
 from pyffi.formats.nif import NifFormat
 
+from io_scene_nif.modules.animation.animation_export import Animation
 from io_scene_nif.modules.object.block_registry import block_store
 from io_scene_nif.modules.property import texture
 from io_scene_nif.modules.property.texture.texture_writer import TextureWriter
@@ -55,6 +56,7 @@ class Texture:
     }
 
     def __init__(self):
+        self.animation_helper = Animation()
         self.dict_mesh_uvlayers = []
         self.texture_writer = TextureWriter()
 
@@ -74,129 +76,6 @@ class Texture:
         for slot in texture_slots:
             used_uvlayers.add(slot.uv_layer)
         return used_uvlayers
-
-    def export_bs_shader_property(self, b_obj=None, b_mat=None):
-        """Export a Bethesda shader property block."""
-        self.determine_texture_types(b_obj, b_mat)
-
-        # create new block
-        if b_obj.niftools_shader.bs_shadertype == 'BSShaderPPLightingProperty':
-            bsshader = NifFormat.BSShaderPPLightingProperty()
-            # set shader options
-            # TODO: FIXME:
-            b_s_type = NifFormat.BSShaderType._enumkeys.index(b_obj.niftools_shader.bsspplp_shaderobjtype)
-            bsshader.shader_type = NifFormat.BSShaderType._enumvalues[b_s_type]
-
-            # Shader Flags
-            if hasattr(bsshader, 'shader_flags'):
-                self.export_shader_flags(b_obj, bsshader)
-
-        if b_obj.niftools_shader.bs_shadertype == 'BSLightingShaderProperty':
-            bsshader = NifFormat.BSLightingShaderProperty()
-            b_s_type = NifFormat.BSLightingShaderPropertyShaderType._enumkeys.index(b_obj.niftools_shader.bslsp_shaderobjtype)
-            bsshader.shader_type = NifFormat.BSLightingShaderPropertyShaderType._enumvalues[b_s_type]
-
-            # UV Offset
-            if hasattr(bsshader, 'uv_offset'):
-                self.export_uv_offset(bsshader)
-
-            # UV Scale
-            if hasattr(bsshader, 'uv_scale'):
-                self.export_uv_scale(bsshader)
-
-            # Texture Clamping mode
-            if not self.base_mtex.texture.image.use_clamp_x:
-                wrap_s = 2
-            else:
-                wrap_s = 0
-            if not self.base_mtex.texture.image.use_clamp_y:
-                wrap_t = 1
-            else:
-                wrap_t = 0
-            bsshader.texture_clamp_mode = (wrap_s + wrap_t)
-
-            # Diffuse color
-            bsshader.skin_tint_color.r = b_mat.diffuse_color.r
-            bsshader.skin_tint_color.g = b_mat.diffuse_color.g
-            bsshader.skin_tint_color.b = b_mat.diffuse_color.b
-            # b_mat.diffuse_intensity = 1.0
-
-            bsshader.lighting_effect_1 = b_mat.niftools.lightingeffect1
-            bsshader.lighting_effect_2 = b_mat.niftools.lightingeffect2
-
-            # Emissive
-            bsshader.emissive_color.r = b_mat.niftools.emissive_color.r
-            bsshader.emissive_color.g = b_mat.niftools.emissive_color.g
-            bsshader.emissive_color.b = b_mat.niftools.emissive_color.b
-            bsshader.emissive_multiple = b_mat.emit
-
-            # gloss
-            bsshader.glossiness = b_mat.specular_hardness
-
-            # Specular color
-            bsshader.specular_color.r = b_mat.specular_color.r
-            bsshader.specular_color.g = b_mat.specular_color.g
-            bsshader.specular_color.b = b_mat.specular_color.b
-            bsshader.specular_strength = b_mat.specular_intensity
-
-            # Alpha
-            if b_mat.use_transparency:
-                bsshader.alpha = (1 - b_mat.alpha)
-
-            # Shader Flags
-            if hasattr(bsshader, 'shader_flags_1'):
-                self.export_shader_flags(b_obj, bsshader)
-
-        if b_obj.niftools_shader.bs_shadertype == 'BSEffectShaderProperty':
-            bsshader = NifFormat.BSEffectShaderProperty()
-
-            # Alpha
-            if b_mat.use_transparency:
-                bsshader.alpha = (1 - b_mat.alpha)
-
-            # clamp Mode
-            bsshader.texture_clamp_mode = 65283
-
-            # Emissive
-            bsshader.emissive_color.r = b_mat.niftools.emissive_color.r
-            bsshader.emissive_color.g = b_mat.niftools.emissive_color.g
-            bsshader.emissive_color.b = b_mat.niftools.emissive_color.b
-            bsshader.emissive_color.a = b_mat.niftools.emissive_alpha
-            bsshader.emissive_multiple = b_mat.emit
-
-            # Shader Flags
-            if hasattr(bsshader, 'shader_flags_1'):
-                self.export_shader_flags(b_obj, bsshader)
-
-        if b_obj.niftools_shader.bs_shadertype == 'None':
-            raise nif_utils.NifError("Export version expected shader. "
-                                     "No shader applied to mesh '%s', these cannot be exported to NIF."
-                                     " Set shader before exporting." % b_obj)
-        # set textures
-        texset = NifFormat.BSShaderTextureSet()
-        bsshader.texture_set = texset
-        if self.base_mtex:
-            texset.textures[0] = self.texture_writer.export_texture_filename(self.base_mtex.texture)
-        if self.normal_mtex:
-            texset.textures[1] = self.texture_writer.export_texture_filename(self.normal_mtex.texture)
-        if self.glow_mtex:
-            texset.textures[2] = self.texture_writer.export_texture_filename(self.glow_mtex.texture)
-        if self.detail_mtex:
-            texset.textures[3] = self.texture_writer.export_texture_filename(self.detail_mtex.texture)
-
-        if b_obj.niftools_shader.bs_shadertype == 'BSLightingShaderProperty':
-            texset.num_textures = 9
-            texset.textures.update_size()
-            if self.detail_mtex:
-                texset.textures[6] = self.texture_writer.export_texture_filename(self.detail_mtex.texture)
-            if self.gloss_mtex:
-                texset.textures[7] = self.texture_writer.export_texture_filename(self.gloss_mtex.texture)
-
-        if b_obj.niftools_shader.bs_shadertype == 'BSEffectShaderProperty':
-            bsshader.source_texture = self.texture_writer.export_texture_filename(self.base_mtex.texture)
-            bsshader.greyscale_texture = self.texture_writer.export_texture_filename(self.glow_mtex.texture)
-
-        return bsshader
 
     def export_texturing_property(self, flags=0x0001, applymode=None, b_mat=None, b_obj=None):
         """Export texturing property."""
@@ -235,7 +114,7 @@ class Texture:
                 pass
             else:
                 # texture slot 0 = base
-                self.animation_helper.texture_animation.export_flip_controller(fliptxt, self.base_mtex.texture, texprop, 0)
+                self.animation_helper.texture.export_flip_controller(fliptxt, self.base_mtex.texture, texprop, 0)
 
         if self.glow_mtex:
             texprop.has_glow_texture = True
@@ -259,7 +138,7 @@ class Texture:
         if self.normal_mtex:
             shadertexdesc = texprop.shader_textures[1]
             shadertexdesc.is_used = True
-            shadertexdesc.texture_data.source = self.texture_writer.export_source_texture(n_texture=self.normal_mtex.texture)
+            shadertexdesc.texture_data.source = TextureWriter.export_source_texture(n_texture=self.normal_mtex.texture)
 
         if self.gloss_mtex:
             if NifOp.props.game not in self.USED_EXTRA_SHADER_TEXTURES:
@@ -270,7 +149,7 @@ class Texture:
             else:
                 shadertexdesc = texprop.shader_textures[2]
                 shadertexdesc.is_used = True
-                shadertexdesc.texture_data.source = self.texture_writer.export_source_texture(n_texture=self.gloss_mtex.texture)
+                shadertexdesc.texture_data.source = TextureWriter.export_source_texture(n_texture=self.gloss_mtex.texture)
 
         if self.dark_mtex:
             texprop.has_dark_texture = True
@@ -292,7 +171,7 @@ class Texture:
             else:
                 shadertexdesc = texprop.shader_textures[3]
                 shadertexdesc.is_used = True
-                shadertexdesc.texture_data.source = self.texture_writer.export_source_texture(n_texture=self.ref_mtex.texture)
+                shadertexdesc.texture_data.source = TextureWriter.export_source_texture(n_texture=self.ref_mtex.texture)
 
     def export_texture_shader_effect(self, tex_prop):
         # export extra shader textures
@@ -309,11 +188,11 @@ class Texture:
             # some texture slots required by the engine
             shadertexdesc_envmap = tex_prop.shader_textures[0]
             shadertexdesc_envmap.is_used = True
-            shadertexdesc_envmap.texture_data.source = self.texture_writer.export_source_texture(filename="RRT_Engine_Env_map.dds")
+            shadertexdesc_envmap.texture_data.source = TextureWriter.export_source_texture(filename="RRT_Engine_Env_map.dds")
 
             shadertexdesc_cubelightmap = tex_prop.shader_textures[4]
             shadertexdesc_cubelightmap.is_used = True
-            shadertexdesc_cubelightmap.texture_data.source = self.texture_writer.export_source_texture(filename="RRT_Cube_Light_map_128.dds")
+            shadertexdesc_cubelightmap.texture_data.source = TextureWriter.export_source_texture(filename="RRT_Cube_Light_map_128.dds")
 
         elif NifOp.props.game == 'CIVILIZATION_IV':
             # some textures end up in the shader texture list there are 4 slots available, so set them up
@@ -336,53 +215,12 @@ class Texture:
         texeff.texture_type = NifFormat.EffectType.EFFECT_ENVIRONMENT_MAP
         texeff.coordinate_generation_type = NifFormat.CoordGenType.CG_SPHERE_MAP
         if b_mat_texslot:
-            texeff.source_texture = self.texture_writer.export_source_texture(b_mat_texslot.texture)
+            texeff.source_texture = TextureWriter.export_source_texture(b_mat_texslot.texture)
             if NifOp.props.game == 'MORROWIND':
                 texeff.num_affected_node_list_pointers += 1
                 texeff.affected_node_list_pointers.update_size()
         texeff.unknown_vector.x = 1.0
-        return self.register_block(texeff)
-
-    def export_uv_offset(self, shader):
-        shader.uv_offset.u = self.base_mtex.offset.x
-        shader.uv_offset.v = self.base_mtex.offset.y
-
-        return shader
-
-    def export_uv_scale(self, shader):
-        shader.uv_scale.u = self.base_mtex.scale.x
-        shader.uv_scale.v = self.base_mtex.scale.y
-
-        return shader
-
-    def export_shader_flags(self, b_obj, shader):
-
-        b_flag_list = b_obj.niftools_shader.bl_rna.properties.keys()
-        if hasattr(shader, 'shader_flags'):
-            for sf_flag in shader.shader_flags._names:
-                if sf_flag in b_flag_list:
-                    b_flag = b_obj.niftools_shader.get(sf_flag)
-                    if b_flag:
-                        sf_flag_index = shader.shader_flags._names.index(sf_flag)
-                        shader.shader_flags._items[sf_flag_index]._value = 1
-
-        if hasattr(shader, 'shader_flags_1'):
-            for sf_flag in shader.shader_flags_1._names:
-                if sf_flag in b_flag_list:
-                    b_flag = b_obj.niftools_shader.get(sf_flag)
-                    if b_flag:
-                        sf_flag_index = shader.shader_flags_1._names.index(sf_flag)
-                        shader.shader_flags_1._items[sf_flag_index]._value = 1
-
-        if hasattr(shader, 'shader_flags_2'):
-            for sf_flag in shader.shader_flags_2._names:
-                if sf_flag in b_flag_list:
-                    b_flag = b_obj.niftools_shader.get(sf_flag)
-                    if b_flag:
-                        sf_flag_index = shader.shader_flags_2._names.index(sf_flag)
-                        shader.shader_flags_2._items[sf_flag_index]._value = 1
-
-        return shader
+        return block_store.register_block(texeff)
 
     def add_shader_integer_extra_datas(self, trishape):
         """Add extra data blocks for shader indices."""
@@ -561,3 +399,55 @@ class Texture:
             else:
                 NifLog.warn("Non-UV texture in mesh '{0}', material '{1}'.\nEither delete all non-UV textures or "
                             "create a UV map for every texture associated with selected object and run the script again.".format(b_obj.name, b_mat.name))
+
+
+def has_diffuse_textures(self, b_mat):
+    if self.b_mat == b_mat:
+        return self.diffusetextures
+
+    for b_mat_texslot in texture.get_used_textslots(b_mat):
+        if b_mat_texslot.use and b_mat_texslot.use_map_color_diffuse:
+            self.diffusetextures.append(b_mat_texslot)
+    return self.diffusetextures
+
+
+def has_glow_textures(self, b_mat):
+    if self.b_mat == b_mat:
+        return self.glowtextures
+
+    for b_mat_texslot in texture.get_used_textslots(b_mat):
+        if b_mat_texslot.use and b_mat_texslot.use_map_emit:
+            self.glowtextures.append(b_mat_texslot)
+    return self.glowtextures
+
+
+def has_bumpmap_textures(self, b_mat):
+    if self.b_mat == b_mat:
+        return self.bumpmaptextures
+
+    for b_mat_texslot in texture.get_used_textslots(b_mat):
+        if b_mat_texslot.use:
+            if b_mat_texslot.texture.use_normal_map is False and b_mat_texslot.use_map_color_diffuse is False:
+                self.bumpmaptextures.append(b_mat_texslot)
+    return self.bumpmaptextures
+
+
+def has_gloss_textures(self, b_mat):
+    if self.b_mat == b_mat:
+        return self.glosstextures
+
+    for b_mat_texslot in texture.get_used_textslots(b_mat):
+        if b_mat_texslot.use and b_mat_texslot.use_map_color_spec:
+            self.glosstextures.append(b_mat_texslot)
+    return self.glosstextures
+
+
+def has_normalmap_textures(self, b_mat):
+    if self.b_mat == b_mat:
+        return self.normalmaptextures
+
+    for b_mat_texslot in texture.get_used_textslots(b_mat):
+        if b_mat_texslot.use:
+            if b_mat_texslot.use_map_color_diffuse is False and b_mat_texslot.texture.use_normal_map and b_mat_texslot.use_map_normal:
+                self.normalmaptextures.append(b_mat_texslot)
+    return self.normalmaptextures
